requirements.txt
adal==1.2.7
azure-common==1.1.27
azure-core==1.17.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.0.0
azure-mgmt-resource==15.0.0
azure-mgmt-storage==16.0.0
azureml==0.2.7
azureml.core==1.37.0
azure.storage.blob==12.9.0
backports.tempfile==1.0
backports.weakref==1.0.post1
biopython==1.79
catboost==0.26
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
docker==4.4.4
graphviz==0.16
idna==3.2
imbalanced-learn==0.8.0
importlib-metadata==4.6.3
isodate==0.6.0
jeepney==0.7.1
jmespath==0.10.0
joblib==1.1.0
jsonpickle==2.0.0
kiwisolver==1.3.1
lightgbm==3.2.1
matplotlib==3.3.4
mock==1.0.1
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
pandas==1.2.0
pathspec==0.9.0
Pillow==8.2.0
plotly==4.14.3
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==227
PyYAML==5.4.1
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
SecretStorage==3.3.1
semantic-version==2.8.5
six==1.15.0
sklearn==0.0
stopwordsiso==0.6.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==1.1.1
xgboost==1.4.0
zipp==3.5.0
nlpaug==1.1.8
azureml-dataset-runtime==1.36.0
wordcloud==1.8.1
torch==1.7.1
imblearn==0.0
transformers==4.12.3
torchtext==0.11.1